+++ example/index.html
@@ -1,6 +1,6 @@
 <html>
   <head>
-    <title>hook-style</title>
+    <title>styled-hooks</title>
     <meta name="viewport" content="width=device-width,initial-scale=1" />
     <link rel="icon" href="/" />
     <style>
